Addressing PR issues
