+++ src/features/commands.ts
@@ -13,7 +13,7 @@ import * as protocol from '../omnisharp/protocol';
 import * as vscode from 'vscode';
 import { DotNetAttachItemsProviderFactory, AttachPicker, RemoteAttachPicker } from './processPicker';
 import { generateAssets } from '../assets';
-import { DebugAdapterDecriptor } from '../coreclr-debug/activate';
+import { ClrDebugDescriptorFactory } from '../coreclr-debug/clrDebugDescriptorFactory';
 import { ShowOmniSharpChannel, CommandDotNetRestoreStart, CommandDotNetRestoreProgress, CommandDotNetRestoreSucceeded, CommandDotNetRestoreFailed } from '../omnisharp/loggingEvents';
 import { EventStream } from '../EventStream';
 import { PlatformInformation } from '../platform';
@@ -47,8 +47,8 @@ export default function registerCommands(server: OmniSharpServer, platformInfo: 
     disposable.add(vscode.commands.registerCommand('csharp.listRemoteProcess', async (args) => RemoteAttachPicker.ShowAttachEntries(args, platformInfo)));
 
     // Register command for adapter executable command.
-    disposable.add(vscode.debug.registerDebugAdapterDescriptorFactory(DebugAdapterDecriptor.CORECLR_DEBUG_TYPE, new DebugAdapterDecriptor(platformInfo, eventStream, packageJSON, extensionPath)));
-    disposable.add(vscode.debug.registerDebugAdapterDescriptorFactory(DebugAdapterDecriptor.CLR_DEBUG_TYPE, new DebugAdapterDecriptor(platformInfo, eventStream, packageJSON, extensionPath)));
+    disposable.add(vscode.debug.registerDebugAdapterDescriptorFactory(ClrDebugDescriptorFactory.CORECLR_DEBUG_TYPE, new ClrDebugDescriptorFactory(platformInfo, eventStream, packageJSON, extensionPath)));
+    disposable.add(vscode.debug.registerDebugAdapterDescriptorFactory(ClrDebugDescriptorFactory.CLR_DEBUG_TYPE, new ClrDebugDescriptorFactory(platformInfo, eventStream, packageJSON, extensionPath)));
     disposable.add(vscode.commands.registerCommand('csharp.reportIssue', async () => reportIssue(vscode, eventStream, getDotnetInfo, platformInfo.isValidPlatformForMono(), optionProvider.GetLatestOptions(), monoResolver)));
 
     return new CompositeDisposable(disposable);
